// app/components/interfaces.tsx
// export interface DragItem {
//     id: string
//     top: number
//     left: number
//     type: string
//     isEditing: boolean
//     content: string
//   }
  
export interface CanvasElement {
    id: string;
    type: string;
    // position: { x: number; y: number };
    top: number;
    left: number;
    content: string;
  }
  